refactor(leetcode/.../two_sum)
Made a module
